Fixes issues with the connection to the position controllers.
The ultrasonic sensor interrupt handler Ultrasonic::HandleInputChange() took too
long, leading to the UART for the position controller not having changed back to
receiving soon enough when an answer of the position controllers was expected.

This was fixed twice: The interrupt handler does less work, and the transmit
delay of the position controllers was increased from 50 to 300 microseconds.