Added parsec_odometry node
It is a replacement of parsec_odometry_relay that should be able to detect if
we lost odometry. In that case, it performs an ICP algorithm to get the
tranform between the last laser message and the laser massage received right
after odometry coming back up to find a correction transform for odometry. This
way, we should be able to get continuous and valid odometry even if we need to
change batteries.